zip file renamming

+++ deploy.sh
@@ -17,7 +17,10 @@ sudo rm -rf /var/www/distfiles
 sudo chown -R www-data:www-data /var/www
 sudo mkdir -p /var/www-cloudsim-auth/machines
 
-sudo mv -f $DIR/boto-useast /var/www-cloudsim-auth
+if [ -f $DIR/boto-useast ]
+then
+    sudo mv -f $DIR/boto-useast /var/www-cloudsim-auth
+fi
 
 cd $DIR/..
 zip -r cloudsim.zip cloudsim
